refactor: extracting configurations for token and hosts to HOC
there were a few places where we were reading the token and the host of the decks
It was also not possible to use the project on localhost to create projects and to read them.

This is changed now as it is possible to open on localhost and use the scratch editor
with the cambrian platform locally.