feat: recreating the costumes on every componentMount. Sync on add and remove
On adding and deleting card we change the specific components.
But on componentDidMount we recreate the whole deck.
The reason for this is that we need an API to understand what has changed for a card
and what should be refreshed.

I know that we are making way to many requests here. I know. But we don't have the API
and developing a proper API will take a day or two. So we do the best we can:

1. Add creating a card we add it to costumes.
2. On deleting a card we delete it from costumes.
3. On updating a card we update it's name, at least, but not the image in the costume
4. On component did mount we refresh all cards