feat: separating deck on smaller components
This is needed for the deck-to-costumes-hoc to maintain
the deck for the whole game

We started by just opening the Deck tab when starting the game
but this required changing the way we load the deck and moving
all to deck-to-costumes-hoc. Otherwise we get into a situation
where after loading we have changed the game and we must save
but saving might enter into a recursion as saving again changes
the costumes as it syncs them with the deck. The game at the end is
again not saved.

Additionally the specs were failing as there were a lot of open
"Are you sure you want to leave the site" messages