feat: sync costumes only to first sprite
If there are other sprites do now.

We sync to sprite with drawableID==1.
This is the smallest change. I have doubts that this is the right api
but there seems to be no other.

We also can not sycn with the sprite called "Cards" as this will bring
all the others games.

Waiting for more requirements to change drawableID==1